requirements.txt
accelerate==1.6.0
bitsandbytes==0.45.5
boto3==1.38.18
deepspeed==0.16.7
mlflow==2.21.2
openai==1.100.2
pandas==2.2.3
peft==0.15.2
pytorch-lightning==2.5.1.post0
rich==14.0.0
sacrebleu==2.5.1
sacrebleu[ja]
tensorboard==2.19.0
tensorboardx==2.6.2.2
torch==2.6.0
torchao==0.11.0
torchtune==0.6.1
transformers==4.52.4
unbabel-comet==2.2.5
